Objects

SyntheticObjects.pollen — Function

pollen(sv = (128, 128, 128), dphi::Float64=0.0, dtheta::Float64=0.0)

Create a 2/3D representation of a pollen grain. default array size is (128, 128, 128) but you can define whether the desired array is 2D or 3D.

Arguments

sv::Tuple: A tuple of three integers representing the size of the volume in which the pollen grain will be created. Default is (128, 128, 128).

dphi::Float64: A float representing the phi angle offset in radians. Default is 0.0.

dtheta::Float64: A float representing the theta angle offset in radians. Default is 0.0.

thickness::Float64: A float representing the thickness of the pollen grain in pixels. Default is 0.8.

intensity::Float64: A float representing the intensity of the pollen grain. Default is 1.0.

filled::Bool: A boolean representing whether the pollen grain is filled. Default is false.

Returns

ret::Array{Float64}: A 3D array representing the pollen grain.

Example

pollen((256, 256, 256), 0.0, 0.0)

SyntheticObjects.object_3D — Function

object_3D([DataType], sz, radius=0.8, center=sz.÷2 .+1; thickness=0.8)

Create a 3D object with a hollow sphere (cell membrane), another hollow sphere (nucleus), a hollow small sphere (vesicle), a filled small sphere (), and a line.

Arguments

DataType: The optional datatype of the output array. Default is Float32.

sz: A vector of integers representing the size of the object. Default is (128, 128, 128).

radius: A float (or tuple) representing the relative radius of the cell sphere.

center: A tuple representing the center of the object.

thickness: A float representing the thickness of the membranes in pixels. Default is 0.8.

SyntheticObjects.filaments — Function

filaments([DataType], sz= (128, 128, 128), rand_offset=0.05, rel_theta=1.0, num_filaments=50, apply_seed=true, thickness=0.8)

Create a 2/3D representation of filaments.

Arguments

DataType: The datatype of the output array. Default is Float32.

sz: A tuple of integers representing the size of the volume into which the filaments will be created. Default is (128, 128, 128).

radius: A tuple of real numbers (or a single real number) representing the relative radius of the volume in which the filaments will be created. Default is 0.8. If a tuple is used, the filamets will be created in a corresponding elliptical volume. Note that the radius is only enforced in the version filaments3D which creates the array rather than adding.

rand_offset: A tuple of real numbers representing the random offsets of the filaments in relation to the size. Default is 0.05.

rel_theta: A real number representing the relative theta range of the filaments. Default is 1.0.

num_filaments: An integer representing the number of filaments to be created. Default is 50.

apply_seed: A boolean representing whether to apply a seed to the random number generator. Default is true.

thickness: A real number representing the thickness of the filaments in pixels. Default is 0.8.

The result is added to the obj input array

Example

# create a 100x100x100 volume with 100 filaments where only the central slice has a random arrangement of filaments
julia> obj = filaments((100,100,100); rel_theta=0, rand_offset=(0.2, 0.2, 0));

# create a 100x100x100 volume with 100 filaments arranged in 3D
julia> obj = filaments((100,100,100));

SyntheticObjects.draw_sphere! — Function

draw_sphere!(arr, radius, center; thickness=0.8, intensity=one(eltype(arr)))

Draw a sphere in a 3D array by adding a Gaussian profile to the array.

#Arguments:

arr::Array: A 3D array to which the sphere will be added.

radius: The radius of the sphere as a number or tuple.

center: The center of the sphere as a tuple. DEFAULT: size(arr).÷2 .+1 which is the (bottom-right) pixel from the center of the array

thickness: The thickness of the sphere. Default is 0.8.

intensity::Float64: The intensity of the sphere. Default is 1.0.

Example

julia> arr = zeros(Float32, (128, 128, 128));
julia> draw_sphere!(arr, 10);

julia> draw_sphere!(arr, (20,30,40), (50,30,80));

SyntheticObjects.draw_line! — Function

draw_line!(arr, start, stop; thickness=0.5, intensity=one(eltype(arr)))

Draw a line in a 3D array by adding a Gaussian profile to the array.

#Arguments:

arr::Array: A 3D array to which the line will be added.

start: The starting point of the line as a tuple.

stop: The stopping point of the line as a tuple.

thickness: The thickness of the line. Default is 0.5.

intensity::Float64: The intensity of the line. Default is 1.0.

SyntheticObjects.filaments! — Function

filaments!(obj; radius = 0.8, rand_offset=0.05, rel_theta=1.0, num_filaments=50, apply_seed=true, thickness=0.8)

Create a 2/3D representation of filaments.

Arguments

obj: A 2/3D array representing the volume into which the filaments will be added.

radius: A tuple of real numbers (or a single real number) representing the relative radius of the volume in which the filaments will be created. Default is 0.8. If a tuple is used, the filamets will be created in a corresponding elliptical volume. Note that the radius is only enforced in the version filaments3D which creates the array rather than adding.

rand_offset: A tuple of real numbers representing the random offsets of the filaments in relation to the size. Default is 0.05.

rel_theta: A real number representing the relative theta range of the filaments. Default is 1.0.

num_filaments: An integer representing the number of filaments to be created. Default is 50.

apply_seed: A boolean representing whether to apply a seed to the random number generator. Default is true.

thickness: A real number representing the thickness of the filaments in pixels. Default is 0.8.

The result is added to the obj input array

Example

# create a 100x100x100 volume with 10 filaments where only the central slice has a random arrangement of filaments
julia> obj = rand(100,100,100); # create an array of random numbers
julia> filaments!(obj; num_filaments=10, rel_theta=0, rand_offset=(0.1,0.1,0), intensity=2.0);

SyntheticObjects.spokes_object — Function

spokes_object(imgSize = (256, 256), numSpikes = 21, continuous = true, makeRound = true)

Generates a 2D or 3D representation of a spokes object.

Arguments

imgSize::Tuple{Int, Int}: A tuple of integers representing the size of the image. Default is (256, 256).

numSpikes::Int: An integer representing the number of spikes. Default is 21.

continuous::Bool: A boolean indicating whether the spokes are continuous. Default is true.

makeRound::Bool: A boolean indicating whether the object is round. Default is true.

Returns

obj2::Array{Float64}: A 2D or 3D array representing the spokes object.

Example

spokes_object((512, 512), 30, false, false)

SyntheticObjects.hollow_sphere! — Function

hollow_sphere(obj, radius=0.8, center=sz.÷2 .+1; thickness=0.8)

Create a 3D representation of a hollow sphere.

Arguments

obj: A 3D array representing the object into which the sphere will be added.

radius: A float representing the radius of the sphere.

thickness: A float representing the thickness of the sphere in pixels. Default is 0.8.

center: A tuple representing the center of the sphere. Default is the center of the object.

Returns

sph::Array{Float64}: A 3D array representing the hollow sphere.

Example

# create a centered sphere of 80% of the object size with a thickness of 0.8 pixels
julia> obj = zeros(Float64, (128, 128, 128));
julia> hollow_sphere(obj, 0.8)

SyntheticObjects.resolution_test — Function

resolution_test(sz = (512, 512, 1); numbers_or_alphabets="alphabets")

Create a 2/3D array of alphabets or numbers with varying font sizes and background levels.

Arguments

sz::Tuple: A tuple of two/three integers representing the size of the volume. Default is (512, 512, 1). Note that the first two elements are multiplied by 10 to defined the array size. The third elements will contain varying alphabets or numbers.

numbers_or_alphabets::String: A string representing whether to use alphabets or numbers. Default is "alphabets".

Returns

A 2/3D array of alphabets or numbers with varying font sizes and background levels.

Example

julia> resolution_test(; sz_each_section=(100, 100), num_slices=1, numbers_or_alphabets="alphabets")

SyntheticObjects.hollow_sphere — Function

hollow_sphere([DataType], sz= (128, 128, 128), radius=0.8, center=sz.÷2 .+1; thickness=0.8)

Create a 3D representation of a hollow sphere.

Arguments

DataType: The optional datatype of the output array. Default is Float32.

sz: A vector of integers representing the size of the sphere.

radius: A float representing the radius of the sphere.

thickness: A float representing the thickness of the sphere in pixels. Default is 0.8.

center: A tuple representing the center of the sphere. Default is the center of the object.

Returns

sph::Array{Float64}: A 3D array representing the hollow sphere.

Example

# create a centered sphere of 80% of the object size with a thickness of 0.8 pixels
julia> hollow_sphere()

SyntheticObjects.annotation_3D! — Function

annotation_3D!(sz=(128,128, 1); numbers_or_alphabets="alphabets", font_size=Float64.(minimum(sz[1:2]))-10.0, bkg=0.9)

Create a 3D array of alphabets or numbers with varying font sizes and background levels.

Arguments

sz::Tuple: A tuple of three integers representing the size of the volume. Default is (128, 128, 1).

numbers_or_alphabets::String: A string representing whether to use alphabets or numbers. Default is "alphabets".

font_size::Float64: A float representing the font size. Default is the minimum of the first two elements of sz minus 10.0.

bkg::Float64: A float representing the background level. Default is 0.9.

Returns

A 3D array of alphabets or numbers with varying font sizes and background levels.

SyntheticObjects.matrix_read — Function

function matrix_read(surface) paint the input surface into a matrix image of the same size to access the pixels.